chore: refactor search-box component
