+++ frontend/src/components/text-input/text-input.tsx
@@ -4,11 +4,16 @@ import './text-input.scss';
 type TextInputProps = {
   name?: string;
   id?: string;
+  className ?: string;
   type?: string;
   onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  ref?: React.MutableRefObject<HTMLInputElement | null>;
   value: any;
+  disabled ?: boolean;
+  readOnly ?: boolean;
 }
 
-export const TextInput: React.FC<TextInputProps> = ({name, id, type = 'text', value, onChange}) => (
-  <input className="text-input" name={name} id={id} type={type} onChange={onChange} value={value} />
+export const TextInput: React.FC<TextInputProps> = ({name, id, type = 'text', value, onChange, className = '', disabled = false, readOnly = false, ref = null}) => (
+  <input className={'text-input ' + className} 
+    name={name} ref={ref} id={id} type={type} onChange={onChange} value={value} disabled={disabled} readOnly={readOnly} />
 );
