+++ frontend/src/css/index.scss
@@ -2,7 +2,6 @@
 @import './base';
 @import './header';
 @import './forms';
-@import './search-box';
 @import './login';
 @import './result-list';
 @import './profile-page';
