+++ frontend/src/css/vars.scss
@@ -1,6 +1,7 @@
 :root {
   --main-background: #222;
   --main-color: #fff;
+  --search-background: #333;
   --second-color: deeppink;
   --third-color: #1f2f5f;
   --disabled-color: #777777;
